requirements.txt
absl-py==2.1.0
accelerate==0.29.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
diffusers==0.27.2
filelock==3.13.3
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
ftfy==6.2.0
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
imutils==0.5.4
itsdangerous==2.1.2
jax==0.4.26
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mediapipe==0.10.11
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
protobuf==3.20.3
psutil==5.9.8
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
sympy==1.12
tokenizers==0.15.2
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
xformers==0.0.25.post1
zipp==3.18.1
